requirements.txt
Flask
mysql-connector-python
Flask-MySQLdb
